BUG: Fix JupyterLab build configuration
Pre-build the labextension with webpack so we can use the webpack rules
required for vtk.js and the custom webpack configuration for itk.js.
This is achieved by whitelisting the modules from being excluded by the
webpack-node-externals package.

Use the Unpkg CDN to provide the WebAssembly and WebWorker files loaded
dynamically.