ENH: Add getters
Getters are designed with the expectation that they will be called twice. On the
first pass the getter is added to a queue and a callback function is set on each.
Once the getter resolves the result is set in the results dict and getters are
ready to be called again to return their result.