ENH: Build a CellWatcher class to monitor cells before running
Intercept execute_requests (which will be all cells queued to run after the
Viewer class has been instantiated. Once the itk_viewer has been created the
queue is processed one request at a time. If a cell contains getters, futures
are created for each and a callback function is set to set the result as they
resolve. Once all of the results are set the cell is run and the queue
continues to be processed.