TEST: Exclude failing cells from notebook tests for now
Improved testing is a current work in progress (as discussed in issue #587).
Until we resolve the outstanding issues no viewer is actually being created, so
the queue that waits for cell completion before progressing causes all
subsequent cells after the first viewer is created to fail. This temporarily
ignores those cells in the CI tests. These changes should be removed after
testing is fixed.